Sergio Tessaris found a problem where the records were not being read in correctly (particularly the first one).
I have rewritten the code that creates the records to try and make it simpler, and to try and make sure it works correctly.

This code SEEMS to work correctly, but I have not had a chance to test it properly.

Frankly, someone (probably me), needs to create some PyUnit tests for this code so we KNOW it works correctly.